refactor: structurized the services folder
