+++ src/Sepes.Infrastructure/Service/SandboxResource/Interface/ISandboxCloudResourceService.cs
