+++ src/Sepes.Infrastructure/Service/SandboxResource/Interface/ISandboxResourceDeleteService.cs
