+++ src/Sepes.Infrastructure/Service/VirtualMachine/Interface/IVirtualMachineLookupService.cs
