+++ src/Sepes.Infrastructure/Service/VirtualMachine/Interface/IVirtualMachineRuleService.cs
