+++ src/Sepes.Infrastructure/Service/VirtualMachine/Interface/IVirtualMachineService.cs
