+++ src/Sepes.Infrastructure/Service/VirtualMachine/VirtualMachineRuleService.cs
